pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "A python fastapi project template with dev containers for use with Codespaces, Docker files including Compose, along some other shit to make life easier for you lazy fucks."
authors = [{ name = "ReAXET", email = "joshua.sturm@outlook.com" }]
dependencies = [
    "loguru>=0.7.2",
    "fastapi[all]>=0.109.2",
    "uvicorn[standard]>=0.27.1",
    "rich>=13.7.0",
    "sqlalchemy>=2.0.27",
    "django>=5.0.2",
    "python-dotenv>=1.0.1",
    "psycopg2-binary>=2.9.9",
    "nba-api>=1.4.1",
    "pandas>=2.2.0",
    "polars>=0.20.9",
    "pyarrow>=15.0.0",
    "seaborn>=0.13.2",
    "matplotlib>=3.8.3",
    "importlib-resources>=6.1.1",
    "streamlit>=1.31.1",
]
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
